+++ robot/rover/science/include/Funnel.h
@@ -8,11 +8,10 @@
 
 #include "Stoppable.h"
 #include "Updatable.h"
-#include <cstdint>
 
 class Funnel : public Stoppable, public Updatable {
     protected:
-        uint8_t funnelPWM;
+        int funnelPWM;
     public:
         virtual void funnel();
         virtual void update(unsigned long) override;
